# server/openapi.yaml
openapi: 3.1.0
info:
  title: QFree API
  version: 1.0.0
  description: |
    Smart healthcare queue management and real-time waiting-time information.

    **Auth:** send `Authorization: Bearer <accessToken>` (15 min). Renew with `POST /auth/refresh`,
    which reads the httpOnly `qf_rt` cookie and rotates it.

    **Errors:** `{ "error": { "code": "QUEUE_FULL", "message": "…", "details": … } }`.

    **Real-time:** Socket.IO on the same origin. Handshake `auth: { token }` (optional).
    Client → server: `queue:watch`, `queue:unwatch`, `queue:watch-staff`, `queue:unwatch-staff`, `org:watch` (arg: id, ack `{ok, error?}`).
    Server → client: `queue:snapshot` (QueueSnapshot), `queue:entry` (EntryView), `queue:staff` (StaffView), `org:queue`, `notification:new`.
servers:
  - url: /api
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
    refreshCookie: { type: apiKey, in: cookie, name: qf_rt }
  parameters:
    id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
...
    page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    pageSize: { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
...
    from: { name: from, in: query, schema: { type: string, format: date }, description: Default 29 days ago }
    to: { name: to, in: query, schema: { type: string, format: date }, description: Default today }
    bucket: { name: bucket, in: query, schema: { type: string, enum: [day, week, month], default: day } }
  responses:
    Error:
      description: Error
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NoContent: { description: No content }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, examples: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, QUEUE_FULL, QUEUE_CLOSED, QUEUE_PAUSED, ALREADY_IN_QUEUE, OUTSIDE_HOURS, DOCTOR_UNAVAILABLE, TOO_MANY_ACTIVE_QUEUES, RATE_LIMITED] }
            message: { type: string }
            details: {}
...
    QueueSnapshot:
      type: object
      description: Public, anonymous-safe queue state (no patient identities).
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        tokenPrefix: { type: string, example: QF }
        status: { type: string, enum: [OPEN, PAUSED, CLOSED] }
        organization: { type: object }
        doctor: { type: [object, 'null'] }
        service: { type: [object, 'null'] }
        sessionDate: { type: [string, 'null'], example: '2026-09-27' }
        currentToken: { type: [string, 'null'], example: QF-024 }
        waitingCount: { type: integer, example: 6 }
        servedCount: { type: integer }
        capacity: { type: integer }
        remainingCapacity: { type: integer }
        avgServiceMinutes: { type: number }
        estimatedWaitMinutes: { type: integer, description: Wait for someone joining now }
        isAcceptingPatients: { type: boolean }
        updatedAt: { type: string, format: date-time }
...
    QueueInput:
      type: object
      required: [organizationId, name]
      properties:
        organizationId: { type: string, format: uuid }
        doctorId: { type: [string, 'null'], format: uuid }
        serviceId: { type: [string, 'null'], format: uuid }
        name: { type: string }
        tokenPrefix: { type: string, pattern: '^[A-Z]{1,4}$', default: QF }
        capacity: { type: integer, minimum: 1, maximum: 5000, default: 100 }
        avgServiceMinutes: { type: number, minimum: 1, maximum: 60, default: 10 }
        approachingThreshold: { type: integer, minimum: 1, maximum: 20, default: 3 }
    Analytics:
      type: object
      properties:
        summary: { type: object, description: 'total, served, cancelled, noShow, skipped, avgWaitMinutes, p90WaitMinutes, avgConsultMinutes, completionRate, utilizationPercent' }
        volume: { type: array, items: { type: object } }
        peakHours: { type: array, items: { type: object, properties: { hour: { type: integer }, joins: { type: integer } } } }
        weekdays: { type: array, items: { type: object } }
        perQueue: { type: array, items: { type: object } }
...
security:
  - bearer: []
paths:
...
  /organizations/{id}:
    get: { tags: [Discovery], security: [], parameters: [{ $ref: '#/components/parameters/id' }], responses: { '200': { description: Profile, hours, doctors, services, live queues }, '404': { $ref: '#/components/responses/Error' } } }
    patch: { tags: [Organizations], summary: Update profile (OWNER/MANAGER), parameters: [{ $ref: '#/components/parameters/id' }], responses: { '200': { description: OK }, '403': { $ref: '#/components/responses/Error' } } }
...
  /queues/{id}:
    parameters: [{ $ref: '#/components/parameters/id' }]
    get: { tags: [Queues], security: [], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/QueueSnapshot' } } } }, '404': { $ref: '#/components/responses/Error' } } }
    patch: { tags: [Queue operations], requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/QueueInput' } } } }, responses: { '200': { description: OK } } }
    delete: { tags: [Queue operations], summary: Archive (must be closed), responses: { '204': { $ref: '#/components/responses/NoContent' }, '409': { $ref: '#/components/responses/Error' } } }
...
  /notifications:
    get: { tags: [Notifications], parameters: [{ $ref: '#/components/parameters/page' }, { $ref: '#/components/parameters/pageSize' }, { name: unread, in: query, schema: { type: string, enum: ['true', 'false'] } }], responses: { '200': { description: Paged notifications plus unread count } } }
...
  /notifications/{id}/read:
    patch: { tags: [Notifications], parameters: [{ $ref: '#/components/parameters/id' }], responses: { '204': { $ref: '#/components/responses/NoContent' } } }

  # ───────── Analytics ─────────
  /analytics/doctor:
    get: { tags: [Analytics], parameters: [{ $ref: '#/components/parameters/from' }, { $ref: '#/components/parameters/to' }, { $ref: '#/components/parameters/bucket' }], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Analytics' } } } } } }
  /analytics/organizations/{id}:
    get: { tags: [Analytics], parameters: [{ $ref: '#/components/parameters/id' }, { $ref: '#/components/parameters/from' }, { $ref: '#/components/parameters/to' }, { $ref: '#/components/parameters/bucket' }], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Analytics' } } } } } }
  /analytics/queues/{id}:
    get: { tags: [Analytics], parameters: [{ $ref: '#/components/parameters/id' }, { $ref: '#/components/parameters/from' }, { $ref: '#/components/parameters/to' }, { $ref: '#/components/parameters/bucket' }], responses: { '200': { description: OK } } }
  /analytics/platform:
    get: { tags: [Analytics], summary: Admin only, parameters: [{ $ref: '#/components/parameters/from' }, { $ref: '#/components/parameters/to' }, { $ref: '#/components/parameters/bucket' }], responses: { '200': { description: OK } } }

  # ───────── Issues ─────────
  /issues:
    post: { tags: [Issues], requestBody: { required: true, content: { application/json: { schema: { type: object, required: [subject, description], properties: { subject: { type: string }, description: { type: string } } } } } }, responses: { '201': { description: Created } } }
...
  /admin/issues/{id}:
    patch: { tags: [Admin], parameters: [{ $ref: '#/components/parameters/id' }], responses: { '200': { description: OK } } }
...